improved server Queque used it in the tcp event sender, also improved README

+++ Rust/server/src/config.rs
@@ -40,6 +40,6 @@ pub struct Connection {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Chunks {
     pub render_distance: u16,
-    pub chunks_per_cycle: u16,
+    pub chunks_per_cycle: u32,
     pub seed: u32,
 }
